requirements.txt
pandas
numpy
faker
supabase
python-dotenv